+++ bods-nextjs/app/publish/account/agent/leave/[pk]/page.tsx
@@ -1,32 +1,29 @@
 'use client';
 
 /**
- * Deactivate or reactivate a member's account (toggle is_active).
+ * Leave an organisation you're acting as an agent for.
+ * Django: `/account/agent/leave/<pk>/`
  */
 
 import { useEffect, useState } from 'react';
 import { useParams, useRouter } from 'next/navigation';
-import { HOSTS } from '@/config/client';
+import { HOSTS, publishAppPath } from '@/config/client';
 import { ProtectedRoute } from '@/components/auth/ProtectedRoute';
 import { Breadcrumbs } from '@/components/shared/Breadcrumbs';
 import { ErrorSummary } from '@/components/shared';
 import { api, getCsrfToken } from '@/lib/api-client';
 
-interface MemberDetail {
+interface AgentInviteDetail {
   id: number;
-  username: string;
-  isActive: boolean;
+  organisationName: string;
 }
 
-function ArchiveMember() {
+function AgentInviteLeave() {
   const params = useParams();
   const router = useRouter();
-  const orgId = params.orgId as string;
-  const userId = params.userId as string;
-  const detailUrl = `/publish/org/${orgId}/manage/${userId}`;
-  const manageUrl = `/publish/org/${orgId}/manage`;
+  const inviteId = params.pk as string;
 
-  const [member, setMember] = useState<MemberDetail | null>(null);
+  const [invite, setInvite] = useState<AgentInviteDetail | null>(null);
   const [isLoading, setIsLoading] = useState(true);
   const [isSubmitting, setIsSubmitting] = useState(false);
   const [error, setError] = useState('');
@@ -35,12 +32,12 @@ function ArchiveMember() {
     let isCancelled = false;
 
     api
-      .get<MemberDetail>(`/api/publish/organisation/members/${userId}/`)
+      .get<AgentInviteDetail>(`/api/publish/agent/invite/${inviteId}/`)
       .then((data) => {
-        if (!isCancelled) setMember(data);
+        if (!isCancelled) setInvite(data);
       })
       .catch(() => {
-        if (!isCancelled) setError('Unable to load this member.');
+        if (!isCancelled) setError('Unable to load this invite.');
       })
       .finally(() => {
         if (!isCancelled) setIsLoading(false);
@@ -49,34 +46,32 @@ function ArchiveMember() {
     return () => {
       isCancelled = true;
     };
-  }, [userId]);
+  }, [inviteId]);
 
   const handleConfirm = async () => {
     setIsSubmitting(true);
     setError('');
 
     try {
-      const response = await fetch(`/api/publish/organisation/members/${userId}/toggle-active/`, {
+      const response = await fetch(`/api/publish/agent/invite/${inviteId}/leave/`, {
         method: 'POST',
         credentials: 'include',
         headers: { 'X-CSRFToken': getCsrfToken() },
       });
 
       if (!response.ok) {
-        setError('Unable to update this member. Please try again.');
+        setError('Unable to leave this organisation. Please try again.');
         setIsSubmitting(false);
         return;
       }
 
-      router.push(detailUrl);
+      router.push(publishAppPath('/account'));
       router.refresh();
     } catch {
-      setError('Unable to update this member. Please try again.');
+      setError('Unable to leave this organisation. Please try again.');
       setIsSubmitting(false);
     }
   };
-
-  const action = member?.isActive ? 'deactivate' : 'reactivate';
 
   return (
     <div className="govuk-width-container">
@@ -85,26 +80,31 @@ function ArchiveMember() {
           items={[
             { label: 'Bus Open Data Service', href: HOSTS.www },
             { label: 'Publish Bus Open Data', href: HOSTS.publish },
-            { label: 'User management', href: manageUrl },
-            { label: 'Deactivate/reactivate', current: true },
+            { label: 'My account', href: publishAppPath('/account') },
+            { label: 'Leave organisation', current: true },
           ]}
         />
 
         <div className="govuk-grid-row">
           <div className="govuk-grid-column-two-thirds">
             {isLoading && <p className="govuk-body">Loading...</p>}
-            {!isLoading && error && <ErrorSummary errors={[error]} summaryId="archive-member-error-title" />}
+            {!isLoading && error && <ErrorSummary errors={[error]} summaryId="agent-leave-error-title" />}
 
-            {!isLoading && member && (
+            {!isLoading && invite && (
               <>
                 <h1 className="govuk-heading-l">
-                  Are you sure you want to {action} {member.username}?
+                  Are you sure you want to leave {invite.organisationName}?
                 </h1>
+                <p className="govuk-body">You will no longer be able to publish or review data for this organisation.</p>
                 <div className="govuk-button-group">
                   <button type="button" className="govuk-button" disabled={isSubmitting} onClick={handleConfirm}>
                     Confirm
                   </button>
-                  <button type="button" className="govuk-button govuk-button--secondary" onClick={() => router.push(detailUrl)}>
+                  <button
+                    type="button"
+                    className="govuk-button govuk-button--secondary"
+                    onClick={() => router.push(publishAppPath('/account'))}
+                  >
                     Cancel
                   </button>
                 </div>
@@ -117,10 +117,10 @@ function ArchiveMember() {
   );
 }
 
-export default function ArchiveMemberPage() {
+export default function AgentInviteLeavePage() {
   return (
     <ProtectedRoute>
-      <ArchiveMember />
+      <AgentInviteLeave />
     </ProtectedRoute>
   );
 }
