+++ bods-nextjs/app/publish/account/manage/[pk]/activate/page.tsx
@@ -1,0 +1,1 @@
+export { default } from '../archive/page';
